+++ crates/board/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Patch
 
+- Update dependencies
 - Remove workaround lint false positive
 
 ## 0.7.0